Split CspApiWeb.JSON back into per-resource modules
The earlier consolidation into a single CspApiWeb.JSON was the wrong
call. `mix phx.gen.json` generates per-resource `*JSON` modules
(JurisdictionJSON, StandardSetJSON, …) that pair 1:1 with controllers,
and the consolidated name shadows Elixir 1.18+'s stdlib JSON module.

Restored:
  CspApiWeb.JurisdictionJSON   summary/1, full/2
  CspApiWeb.StandardSetJSON    summary/1, full/1
  CspApiWeb.UserJSON           show/1
  CspApiWeb.ActivityJSON       show/1
  CspApiWeb.PullRequestJSON    full/1, summary/1

Controllers updated to alias and call the per-resource modules.
