+++ elixir/lib/csp_api_web/controllers/users_controller.ex
@@ -2,11 +2,11 @@ defmodule CspApiWeb.UsersController do
   use CspApiWeb, :controller
 
   alias CspApi.{Users, PullRequests}
-  alias CspApiWeb.JSON
+  alias CspApiWeb.UserJSON
 
   def signed_in(conn, params) do
     user = Users.upsert_signed_in(params)
-    json(conn, %{data: JSON.user(user)})
+    json(conn, %{data: UserJSON.show(user)})
   end
 
   def show(conn, %{"email" => email}) do
@@ -16,14 +16,14 @@ defmodule CspApiWeb.UsersController do
 
       user ->
         prs = PullRequests.list_for_user(user.id)
-        json(conn, %{data: JSON.user(Map.put(user, :pullRequests, prs))})
+        json(conn, %{data: UserJSON.show(Map.put(user, :pullRequests, prs))})
     end
   end
 
   def set_allowed_origins(conn, %{"id" => id, "data" => data}) when is_list(data) do
     case Users.set_allowed_origins(id, data) do
       nil -> conn |> put_status(:not_found) |> json(%{error: "User not found"})
-      user -> json(conn, %{data: JSON.user(user)})
+      user -> json(conn, %{data: UserJSON.show(user)})
     end
   end
 end
